v0.3.0
Vastly improved logic for Bucket Sort. Works now as intended in a 3D array instead of a 1D with funky distribution.
Furthermore, added the functionality to crop black bars as well as a scale to 720p to ease use of drive space, as well as analysis time.
Finally removed Hardware Accelerated parameters for FFmpeg. CPU only is about twice as fast as with the GPU enabled.